Student Page Update
Added: Register for courses functionality, as well as only showing courses based off a users major. View my courses which shows you all the courses a users is currently enrolled in as well as showing the courses a user has dropped. A choose major which allows a user to choose a new major. Choosing a new major will add to your  major because a student can declare multiple majors. No delete major.